+++ src/text/inline.rs
@@ -107,6 +107,7 @@ fn push_values<'s, T: DiffableStr + ?Sized>(
 ///
 /// This is like [`Change`] but with inline highlight info.
 #[derive(Debug, PartialEq, Eq, Hash, Clone, Ord, PartialOrd)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize))]
 pub struct InlineChange<'s, T: DiffableStr + ?Sized> {
     tag: ChangeTag,
     old_index: Option<usize>,
@@ -316,4 +317,21 @@ fn test_line_ops_inline() {
         .flat_map(|op| diff.iter_inline_changes(op))
         .collect::<Vec<_>>();
     insta::assert_debug_snapshot!(&changes);
+}
+
+#[test]
+#[cfg(feature = "serde")]
+fn test_serde() {
+    let diff = TextDiff::from_lines(
+        "Hello World\nsome stuff here\nsome more stuff here\n\nAha stuff here\nand more stuff",
+        "Stuff\nHello World\nsome amazing stuff here\nsome more stuff here\n",
+    );
+    assert_eq!(diff.newline_terminated(), true);
+    let changes = diff
+        .ops()
+        .iter()
+        .flat_map(|op| diff.iter_inline_changes(op))
+        .collect::<Vec<_>>();
+    let json = serde_json::to_string_pretty(&changes).unwrap();
+    insta::assert_snapshot!(&json);
 }